note types can be used to positively filter OMOP CDM pulls
Two new properties in the database_connections.properties file
can be set to specify the table column for NOTE_TYPEs and the
list of values of NOTE_TYPEs to include in the database pull.

A little bit of whitespace tidying up as well.

Normalized a few "physical examination/status" section header
names that had be set to "Unknown".